requirements.txt
google-genai
flask
flask-cors
requests
python-dotenv